Refactor S3 delivery stream into Firehose namespace
