+++ packages/punchcard/lib/stream/collector.ts
@@ -1,23 +1,23 @@
 import core = require('@aws-cdk/core');
 
 import { Type } from '../shape/types/type';
-import { S3DeliveryStreamCollector, S3DeliveryStreamForType } from './delivery-stream';
+import { Firehose } from './delivery-stream';
 import { Kinesis } from './kinesis';
 import { SNS } from './sns';
 import { SQS } from './sqs';
 import { Stream } from './stream';
 
 /**
- * Collects data from an `Enumerable`.
+ * Collects data from an `Stream`.
  *
  * @param T type of collected result
- * @param E source enumerable
+ * @param E source Stream
  */
-export interface Collector<T, E extends Stream<any, any, any, any>> {
+export interface Collector<T, S extends Stream<any, any, any, any>> {
   /**
-   * Create constructs to collect data from the enumerable an
+   * Create Constructs to collect data from the `Stream` and returns the result of that collection.
    */
-  collect(scope: core.Construct, id: string, enumerable: E): T
+  collect(scope: core.Construct, id: string, stream: S): T
 }
 
 /**
@@ -25,7 +25,7 @@ export interface Collector<T, E extends Stream<any, any, any, any>> {
  */
 export namespace Collectors {
   /**
-   * Collects data from an `Enumerable` into a new SQS Queue.
+   * Collects data from an `Stream` into a new SQS Queue.
    *
    * @param props queue properties
    */
@@ -34,7 +34,7 @@ export namespace Collectors {
   }
 
   /**
-   * Collects data from an `Enumerable` into a new Kinesis Stream.
+   * Collects data from an `Stream` into a new Kinesis Stream.
    *
    * @param props stream properties
    */
@@ -43,16 +43,16 @@ export namespace Collectors {
   }
 
   /**
-   * Collects data from an `Enumerable` into S3 via a Firehose Delivery Stream.
+   * Collects data from an `Stream` into S3 via a Firehose Delivery Stream.
    *
    * @param props stream properties
    */
-  export function toS3DeliveryStream<T extends Type<any>>(props: S3DeliveryStreamForType<T>): S3DeliveryStreamCollector<T, any> {
-    return new S3DeliveryStreamCollector<T, any>(props);
+  export function toS3DeliveryStream<T extends Type<any>>(props: Firehose.S3DeliveryStreamForType<T>): Firehose.S3DeliveryStreamCollector<T, any> {
+    return new Firehose.S3DeliveryStreamCollector<T, any>(props);
   }
 
   /**
-   * Collects data from an `Enumerable` into a new SNS Topic.
+   * Collects data from an `Stream` into a new SNS Topic.
    *
    * @param props topic properties
    */
