+++ packages/punchcard/lib/stream/glue.ts
@@ -11,17 +11,17 @@ import { Collector } from './collector';
 import { DependencyType, EventType, Stream } from './stream';
 
 /**
- * Creates a new Glue `Table` and publishes data from an enumerable to it.
+ * Creates a new Glue `Table` and publishes data from a `Stream` to it.
  *
  * @typeparam T type of notififcations sent to (and emitted from) the Glue Table.
  */
 export class GlueTableCollector<S extends Shape, P extends Glue.Partition, E extends Stream<any, RuntimeShape<S>, any, any>> implements Collector<CollectedGlueTable<S, P, E>, E> {
   constructor(private readonly props: Glue.TableProps<S, P>) { }
 
-  public collect(scope: core.Construct, id: string, enumerable: E): CollectedGlueTable<S, P, E> {
+  public collect(scope: core.Construct, id: string, stream: E): CollectedGlueTable<S, P, E> {
     return new CollectedGlueTable(scope, id, {
       ...this.props,
-      enumerable
+      stream
     });
   }
 }
@@ -29,24 +29,26 @@ export class GlueTableCollector<S extends Shape, P extends Glue.Partition, E ext
 /**
  * Properties for creating a collected `Table`.
  */
-export interface CollectedGlueTableProps<S extends Shape, P extends Glue.Partition, E extends Stream<any, RuntimeShape<S>, any, any>> extends Glue.TableProps<S, P> {
+export interface CollectedGlueTableProps<T extends Shape, P extends Glue.Partition, S extends Stream<any, RuntimeShape<T>, any, any>> extends Glue.TableProps<T, P> {
   /**
-   * Source of the data; an enumerable.
+   * Source of the data; a stream.
    */
-  readonly enumerable: E;
+  readonly stream: S;
 }
 
 /**
- * A Glue `Table` produced by collecting data from an `Enumerable`.
+ * A Glue `Table` produced by collecting data from a `Stream`.
  *
- * @typeparam T type of notififcations sent to, and emitted from, the Glue Table.
+ * @typeparam T shape of data
+ * @typeparam P shape of partition keys
+ * @typeparam S stream of data to ingest into the table
  */
-export class CollectedGlueTable<S extends Shape, P extends Glue.Partition, E extends Stream<any, any, any, any>> extends Glue.Table<S, P> {
-  public readonly sender: Function<EventType<E>, void, Dependency.List<Cons<DependencyType<E>, Dependency<Glue.Table.Client<S, P>>>>>;
+export class CollectedGlueTable<T extends Shape, P extends Glue.Partition, S extends Stream<any, any, any, any>> extends Glue.Table<T, P> {
+  public readonly sender: Function<EventType<S>, void, Dependency.List<Cons<DependencyType<S>, Dependency<Glue.Table.Client<T, P>>>>>;
 
-  constructor(scope: core.Construct, id: string, props: CollectedGlueTableProps<S, P, E>) {
+  constructor(scope: core.Construct, id: string, props: CollectedGlueTableProps<T, P, S>) {
     super(scope, id, props);
-    this.sender = props.enumerable.forBatch(this.resource, 'ToTable', {
+    this.sender = props.stream.forBatch(this.resource, 'ToTable', {
       depends: this,
       handle: async (events, self) => {
         self.sink(events);
@@ -56,20 +58,19 @@ export class CollectedGlueTable<S extends Shape, P extends Glue.Partition, E ext
 }
 
 /**
- * Add a utility method `toGlueTable` for `Enumerable` which uses the `TableCollector` to produce Glue `Tables`.
+ * Add a utility method `toGlueTable` for `Stream` which uses the `TableCollector` to produce Glue `Tables`.
  */
 declare module './stream' {
-  interface Stream<E, T, D extends any[], R extends StreamRuntime> {
+  interface Stream<E, T, D extends any[], R extends Stream.Config> {
     /**
      * Collect data to S3 via a Firehose Delivery Stream.
      *
      * @param scope
      * @param id
      * @param tableProps properties of the created s3 delivery stream
-     * @param runtimeProps optional runtime properties to configure the function processing the enumerable's data.
-     * @typeparam T concrete type of data flowing to s3
+     * @param runtimeConfig optional runtime properties to configure the function processing the stream's data.
      */
-    toGlueTable<S extends Shape, T extends StructType<S> & Type<T>, P extends Glue.Partition>(scope: core.Construct, id: string, tableProps: Glue.TableProps<S, P>, runtimeProps?: R): CollectedGlueTable<S, P, this>;
+    toGlueTable<S extends Shape, T extends StructType<S>, P extends Glue.Partition>(scope: core.Construct, id: string, tableProps: Glue.TableProps<S, P>, runtimeConfig?: R): CollectedGlueTable<S, P, this>;
   }
 }
 Stream.prototype.toGlueTable = function(scope: core.Construct, id: string, tableProps: any, runtimeProps?: any): any {
