+++ packages/punchcard/lib/stream/stream.ts
@@ -9,16 +9,6 @@ export type DataType<E extends Stream<any, any, any, any>> = E extends Stream<an
 export type DependencyType<E extends Stream<any, any, any, any>> = E extends Stream<any, any, infer D, any> ? D : never;
 
 /**
- * Props to configure an `Enumerable's` evaluation runtime properties.
- */
-export interface StreamRuntime {
-  /**
-   * The executor service of a `Enumerable` can always be customized.
-   */
-  executorService?: Lambda.ExecutorService;
-}
-
-/**
  * Represents chainable async operations on a cloud data structure.
  *
  * @typeparam E type of event that triggers the computation, i.e. SQSEvent to a Lambda Function
@@ -26,7 +16,7 @@ export interface StreamRuntime {
  * @typeparam D runtime dependencies
  * @typeparam R runtime configuration
  */
-export abstract class Stream<E, T, D extends any[], R extends StreamRuntime> {
+export abstract class Stream<E, T, D extends any[], R extends Stream.Config> {
   constructor(
       protected readonly previous: Stream<E, any, any, R>,
       protected readonly f: (value: AsyncIterableIterator<any>, clients: Clients<D>) => AsyncIterableIterator<T>,
@@ -126,7 +116,7 @@ export abstract class Stream<E, T, D extends any[], R extends StreamRuntime> {
     handle: (value: T, deps: Client<D2>) => Promise<any>;
     props?: R;
   }): Lambda.Function<E, any, D2 extends undefined ? Dependency.List<D> : Dependency.List<Cons<D, D2>>> {
-    // TODO: let the enumerable type determine default executor service
+    // TODO: let the stream type determine default executor service
     const executorService = (input.props && input.props.executorService) || new Lambda.ExecutorService({
       memorySize: 128,
       timeout: core.Duration.seconds(10)
@@ -201,5 +191,17 @@ export abstract class Stream<E, T, D extends any[], R extends StreamRuntime> {
    */
   public collect<T>(scope: core.Construct, id: string, collector: Collector<T, this>): T {
     return collector.collect(scope, id, this);
+  }
+}
+
+export namespace Stream {
+  /**
+   * Props to configure a `Stream's` evaluation runtime properties.
+   */
+  export interface Config {
+    /**
+     * The executor service of a `Stream` can always be customized.
+     */
+    executorService?: Lambda.ExecutorService;
   }
 }
